Move SuperblockPtr from quorum.h to superblock.h
This prepares for removal of superblocks (claims) from CBlock and helps
to improve compilation performance.
